move and rename components/ActiveRecord to contexts/ActiveRecordProvider
